+++ packages/dumb/src/utils/datetime/datetime.util.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -28,7 +28,7 @@
 import _ from 'lodash'
 import moment from 'moment-timezone'
 import i18next from 'i18next'
-import { DurationUnit } from 'medical-domain'
+import { DurationUnit, TimePrefs } from 'medical-domain'
 
 const t = i18next.t.bind(i18next)
 
@@ -36,6 +36,7 @@ export const THIRTY_MINS = 1800000
 export const ONE_HR = 3600000
 export const THREE_HRS = 10800000
 export const TWENTY_FOUR_HRS = 86400000
+export const TIMEZONE_UTC = 'UTC'
 
 /**
  * getHourMinuteFormat
@@ -46,61 +47,11 @@ export const getHourMinuteFormat = (): string => {
 }
 
 /**
- * getHourMinuteFormatNoSpace
- * @returns string according to translation
- */
-export const getHourMinuteFormatNoSpace = (): string => {
-  return t('h:mma')
-}
-
-/**
- * getSimpleHourFormat
- * @returns string according to translation
- */
-export const getSimpleHourFormat = (): string => {
-  return t('ha')
-}
-
-/**
  * getDayFormat
  * @returns string according to translation
  */
 export const getDayFormat = (): string => {
   return t('dddd, MMMM D')
-}
-
-/**
- * getLongDayFormat
- * @returns string according to translation
- */
-export const getLongDayFormat = (): string => {
-  return t('MMM D, YYYY')
-}
-
-/**
- * getSimpleHourFormatSpace
- * @returns string according to translation
- */
-export const getSimpleHourFormatSpace = (): string => {
-  return t('h a')
-}
-
-/**
- * getLongFormat
- * @returns string according to translation
- */
-export const getLongFormat = (): string => {
-  return t('ddd, MMM D, Y')
-}
-
-/**
- * Moment format for:
- * - English: 'MMM D, YYYY h:mm a'
- * - French: 'D MMM YYYY, H:mm'
- * @returns {string} Format according to translation
- */
-export const getLongDayHourFormat = (): string => {
-  return t('MMM D, YYYY h:mm a')
 }
 
 /**
@@ -129,66 +80,20 @@ export const getBrowserTimezone = (): string => {
  *
  * @return {String} timezoneName from timePrefs, browser, or fallback to 'UTC'
  */
-export const getTimezoneFromTimePrefs = (timePrefs: { timezoneAware: boolean, timezoneName: string }): string => {
+export const getTimezoneFromTimePrefs = (timePrefs: TimePrefs): string => {
   if (typeof timePrefs === 'object' && !_.isEmpty(timePrefs)) {
-    return timePrefs.timezoneName ?? getBrowserTimezone() ?? 'UTC'
+    return timePrefs.timezoneName ?? getBrowserTimezone() ?? TIMEZONE_UTC
   }
-  return 'UTC'
-}
-
-/**
- * formatBirthdate
- * @param {Object} patient - Tidepool patient object containing profile
- *
- * @return {String} formatted birthdate, e.g., 'Jul 4, 1975'; empty String if none found
- */
-export const formatBirthdate = (patient: { profile: { patient: { birthday: string } } }): string => {
-  const birthday = _.get(patient, 'profile.patient.birthday', '')
-  if (birthday) {
-    return moment.utc(birthday).format(t('birthday-format'))
-  }
-  return ''
-}
-
-/**
- * formatClocktimeFromMsPer24
- * @param {number} milliseconds - positive integer representing a time of day
- *                            in milliseconds within a 24-hr day
- * @param {string} format - optional moment display format string; default is 'h:mm a'
- *
- * @return {string} formatted clocktime, e.g., '12:05 pm'
- */
-export const formatClocktimeFromMsPer24 = (milliseconds: number, format = getHourMinuteFormat()): string => {
-  if (!Number.isFinite(milliseconds) || milliseconds < 0 || milliseconds > TWENTY_FOUR_HRS) {
-    throw new Error('First argument must be a value in milliseconds per twenty-four hour day!')
-  }
-  return moment.utc(milliseconds).format(format)
-}
-
-/**
- * formatDateRange
- * @param {string} startDate - A moment-compatible date object or string
- * @param {string} endDate - A moment-compatible date object or string
- * @param {string} format - Optional. The moment format string to parse startDate and endDate with
- * @param {string} timezone Default to UTC
- */
-export const formatDateRange = (startDate: string, endDate: string, format: string, timezone = 'UTC'): string => {
-  const start = moment.tz(startDate, format, timezone)
-  const end = moment.tz(endDate, format, timezone)
-
-  const isSameYear = start.isSame(end, 'year')
-  const startFormat = isSameYear ? start.format(t('MMM D')) : start.format(t('MMM D, YYYY'))
-  const endFormat = end.format(t('MMM D, YYYY'))
-  return `${startFormat} - ${endFormat}`
+  return TIMEZONE_UTC
 }
 
 /**
  * Format a duration
  * @param {number} duration - positive integer duration in milliseconds
- * @param {{condensed?: boolean}} opts - options
+ * @param {boolean} condensed - whether the return string should be condensed (e.g: '1min' or '1m')
  * @return {string} formatted duration, e.g., '1¼ h'
  */
-export const formatDuration = (duration: number, opts: { condensed: boolean }): string => {
+export const formatDuration = (duration: number, condensed: boolean = false): string => {
   const momentDuration = moment.duration(duration)
   const days = momentDuration.days()
   const hours = momentDuration.hours()
@@ -201,7 +106,7 @@ export const formatDuration = (duration: number, opts: { condensed: boolean }): 
   const TWO_THIRDS = '⅔'
   const THREE_QUARTERS = '¾'
 
-  if (opts.condensed) {
+  if (condensed) {
     const formatted = {
       days: '',
       hours: '',
@@ -268,61 +173,9 @@ export const formatDuration = (duration: number, opts: { condensed: boolean }): 
  *
  * @return {String} formatted datetime, e.g., 'Sunday, January 1'
  */
-export const formatLocalizedFromUTC = (utc: string | number | Date | moment.Moment, timePrefs: { timezoneAware: boolean, timezoneName: string }, format = getDayFormat()): string => {
+export const formatLocalizedFromUTC = (utc: string | number | Date | moment.Moment, timePrefs: TimePrefs, format = getDayFormat()): string => {
   const timezone = getTimezoneFromTimePrefs(timePrefs)
   return moment.utc(utc).tz(timezone).format(format)
-}
-
-/**
- * getHammertimeFromDatumWithTimePrefs
- * @param {Object} datum - a Tidepool datum with a normalTime (required) and deviceTime (optional)
- * @param {Object} timePrefs - object containing timezoneAware Boolean and timezoneName String
- *
- * @return {Number} Integer hammertime (i.e., UTC time in milliseconds)
- */
-export const getHammertimeFromDatumWithTimePrefs = (datum: { normalTime: string, deviceTime: string }, timePrefs: { timezoneAware: boolean, timezoneName: string }): number | null => {
-  let hammertime
-  if (timePrefs.timezoneAware) {
-    if (!_.isUndefined(datum.normalTime)) {
-      hammertime = Date.parse(datum.normalTime)
-    } else {
-      hammertime = null
-    }
-  } else if (!_.isUndefined(datum.deviceTime)) {
-    hammertime = Date.parse(datum.deviceTime)
-  } else {
-    hammertime = null
-  }
-  if (_.isNaN(hammertime)) {
-    throw new Error(
-      'Check your input datum; could not parse `normalTime` or `deviceTime` with Date.parse.'
-    )
-  }
-  return hammertime
-}
-
-/**
- * getLocalizedCeiling
- * @param {string} utc - Zulu timestamp (Integer hammertime also OK)
- * @param {object} timePrefs - object containing timezoneAware Boolean and timezoneName String
- *
- * @return {Date} a JavaScript Date, the closest (future) midnight according to timePrefs;
- *                  if utc is already local midnight, returns utc
- */
-export const getLocalizedCeiling = (utc: string | number | Date | moment.Moment, timePrefs: { timezoneAware: boolean, timezoneName: string }): Date => {
-  if (utc instanceof Date) {
-    throw new Error('`utc` must be a ISO-formatted String timestamp or integer hammertime!')
-  }
-  const timezone = getTimezoneFromTimePrefs(timePrefs)
-  const startOfDay = moment.utc(utc)
-    .tz(timezone)
-    .startOf('day')
-
-  const utcHammertime = (typeof utc === 'string') ? Date.parse(utc) : utc
-  if (startOfDay.valueOf() === utcHammertime) {
-    return startOfDay.toDate()
-  }
-  return startOfDay.add(1, 'day').toDate()
 }
 
 export const convertValueToMinutes = (durationValue: number, durationUnit: DurationUnit): number => {